+++ src/main/java/edu/ie3/datamodel/models/input/system/type/HpTypeInput.java
@@ -8,6 +8,7 @@ package edu.ie3.datamodel.models.input.system.type;
 import edu.ie3.datamodel.models.StandardUnits;
 import edu.ie3.util.quantities.interfaces.Currency;
 import edu.ie3.util.quantities.interfaces.EnergyPrice;
+import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
 import javax.measure.quantity.Power;
@@ -38,6 +39,31 @@ public class HpTypeInput extends SystemParticipantTypeInput {
       ComparableQuantity<Power> pThermal) {
     super(uuid, id, capex, opex, sRated.to(StandardUnits.S_RATED), cosPhiRated);
     this.pThermal = pThermal.to(StandardUnits.ACTIVE_POWER_IN);
+  }
+
+  /**
+   * @param uuid of the input entity
+   * @param id of this type of HP
+   * @param capex Capital expense for this type of HP (typically in €)
+   * @param opex Operating expense for this type of HP (typically in €)
+   * @param cosPhiRated Power factor for this type of HP
+   * @param sRated Rated apparent power
+   * @param pThermal Thermal output of the heat pump, when sRated * cosPhiRated is consumed
+   *     electrically
+   * @param additionalInformation Of the input
+   */
+  public HpTypeInput(
+      UUID uuid,
+      String id,
+      ComparableQuantity<Currency> capex,
+      ComparableQuantity<EnergyPrice> opex,
+      ComparableQuantity<Power> sRated,
+      double cosPhiRated,
+      ComparableQuantity<Power> pThermal,
+      Map<String, String> additionalInformation) {
+    super(uuid, id, capex, opex, sRated.to(StandardUnits.S_RATED), cosPhiRated);
+    this.pThermal = pThermal.to(StandardUnits.ACTIVE_POWER_IN);
+    setAdditionalInformation(additionalInformation);
   }
 
   public ComparableQuantity<Power> getpThermal() {
@@ -79,6 +105,8 @@ public class HpTypeInput extends SystemParticipantTypeInput {
         + getCosPhiRated()
         + "pThermal="
         + pThermal
+        + ", additionalInformation="
+        + getAdditionalInformation()
         + '}';
   }